% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testFile1.R
\name{testFunc1}
\alias{testFunc1}
\title{Test File #1}
\usage{
testFunc1(x, y)
}
\arguments{
\item{x}{Integer or real.}

\item{y}{Integer or real.}
}
\description{
This test file serves as a demonstration of the `@tests` tag infrastructure.
It contains multiple `@tests` blocks and uses both `tinytest` and `testthat`.
}
\examples{
res <- testFunc1(3, 5)

}
